feat(workers) refactored workers and removed repeated code
